Move breakdown refs into useContracts composable
